requirements.txt
langchain>=0.1.20
langchain-community>=0.0.38
langchain-core>=0.1.52
python-dotenv
langchain-ollama==0.3.6
google-search-results
fastapi
uvicorn[standard]
